+++ R/add_locked_out_constraints.R
@@ -357,14 +357,14 @@ methods::setMethod("add_locked_out_constraints",
       is_same_crs(x$data$cost, locked_out),
       msg = paste(
         "{.arg locked_out} and planning units for {.arg x}",
-        "must have the same coordinate reference system"
+        "must have the same coordinate reference system."
       )
     )
     assert(
       is_spatial_extents_overlap(x$data$cost, locked_out),
       msg = paste(
         "{.arg locked_out} and planning units for {.arg x}",
-        "must have overlapping spatial extents"
+        "must have overlapping spatial extents."
       )
     )
     # add constraints
@@ -399,14 +399,14 @@ methods::setMethod("add_locked_out_constraints",
       is_same_crs(x$data$cost, locked_out),
       msg = paste(
         "{.arg locked_out} and planning units for {.arg x}",
-        "must have the same coordinate reference system"
+        "must have the same coordinate reference system."
       )
     )
     assert(
       is_spatial_extents_overlap(x$data$cost, locked_out),
       msg = paste(
         "{.arg locked_out} and planning units for {.arg x}",
-        "must have overlapping spatial extents"
+        "must have overlapping spatial extents."
       )
     )
     # create matrix with statuses