+++ R/st_geometry_classes.R
@@ -12,5 +12,19 @@ NULL
 #' @noRd
 st_geometry_classes <- function(x) {
   assert(inherits(x, "sf"))
-  vapply(sf::st_geometry(x), class, character(3))[2, ]
+  if (inherits(sf::st_geometry(x), "sfc_POLYGON")) {
+    return(rep("POLYGON", nrow(x)))
+  } else if (inherits(sf::st_geometry(x), "sfc_POINT")) {
+    return(rep("POINT", nrow(x)))
+  } else if (inherits(sf::st_geometry(x), "sfc_MULTIPOLYGON")) {
+    return(rep("MULTIPOLYGON", nrow(x)))
+  } else if (inherits(sf::st_geometry(x), "sfc_MULTIPOINT")) {
+    return(rep("MULTIPOINT", nrow(x)))
+  } else if (inherits(sf::st_geometry(x), "sfc_LINESTRING")) {
+    return(rep("LINESTRING", nrow(x)))
+  } else if (inherits(sf::st_geometry(x), "sfc_MULTILINESTRING")) {
+    return(rep("MULTILINESTRING", nrow(x)))
+  } else {
+    vapply(sf::st_geometry(x), function(y) class(y)[[2]], character(1))
+  }
 }